Split ArtBoard classes into their own files
