+++ art/src/main/java/vwaber/copixelate/art/BitmapData.kt
@@ -1,0 +1,6 @@
+package vwaber.copixelate.art
+
+class BitmapData(
+    val size: Point,
+    val pixels: IntArray
+)
